HDX-9781 Move dataset VAT
