+++ src/hapi_schema/db_views_as_tables.py
@@ -46,3 +46,12 @@ class DBAdmin2VAT(Base):
     admin1_is_unspecified: Mapped[bool] = mapped_column(Boolean)
     location_code: Mapped[str] = mapped_column(String(128), index=True)
     location_name: Mapped[str] = mapped_column(String(512), index=True)
+
+
+class DBDatasetVAT(Base):
+    __tablename__ = "dataset_vat"
+    hdx_id: Mapped[str] = mapped_column(String(36), primary_key=True)
+    hdx_stub: Mapped[str] = mapped_column(String(128), index=True)
+    title: Mapped[str] = mapped_column(String(1024))
+    hdx_provider_stub: Mapped[str] = mapped_column(String(128), index=True)
+    hdx_provider_name: Mapped[str] = mapped_column(String(512), index=True)
